Refactor module ams.utils for simplicity
